Carter

+++ db/migrate/20150711164928_drop_bakset.rb
@@ -1,5 +1,7 @@
 class DropBakset < ActiveRecord::Migration
   def change
-  	drop_table :baksets
+    if ActiveRecord::Base.connection.table_exists? "baksets"
+      drop_table :baksets
+    end
   end
 end
